LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/update-bytecheck.https-expected.txt


FAIL Test(cors: false, main: default, imported: default) assert_array_equals: property 2, expected null but got object "[object ServiceWorker]"
FAIL Test(cors: false, main: default, imported: time) promise_test: Unhandled rejection with value: object "TypeError: Cannot update a service worker with a requested script URL whose newest worker has a different script URL"
FAIL Test(cors: false, main: time, imported: default) promise_test: Unhandled rejection with value: object "TypeError: Cannot update a service worker with a requested script URL whose newest worker has a different script URL"
FAIL Test(cors: false, main: time, imported: time) promise_test: Unhandled rejection with value: object "TypeError: Cannot update a service worker with a requested script URL whose newest worker has a different script URL"
FAIL Test(cors: true, main: default, imported: default) promise_test: Unhandled rejection with value: object "TypeError: Cannot update a service worker with a requested script URL whose newest worker has a different script URL"
FAIL Test(cors: true, main: default, imported: time) promise_test: Unhandled rejection with value: object "TypeError: Cannot update a service worker with a requested script URL whose newest worker has a different script URL"
FAIL Test(cors: true, main: time, imported: default) promise_test: Unhandled rejection with value: object "TypeError: Cannot update a service worker with a requested script URL whose newest worker has a different script URL"
FAIL Test(cors: true, main: time, imported: time) promise_test: Unhandled rejection with value: object "TypeError: Cannot update a service worker with a requested script URL whose newest worker has a different script URL"
